refactor : useDraft, useModalControl,useWorkData 로 커스텀 훅 분리를 통한 유지보수 및 재활용성 향상
